refactor: rename swap2 to swap
