+++ apps/cowswap-frontend/src/modules/swap/containers/Warnings/index.tsx
