+++ apps/cowswap-frontend/src/modules/trade/state/derivedTradeStateAtom.ts
@@ -7,11 +7,11 @@ import { atom } from 'jotai'
  */
 import { advancedOrdersDerivedStateAtom } from 'modules/advancedOrders/state/advancedOrdersAtom'
 import { limitOrdersDerivedStateAtom } from 'modules/limitOrders/state/limitOrdersRawStateAtom'
-import { swapDerivedStateAtom } from 'modules/swap2/state/swapRawStateAtom'
 import { yieldDerivedStateAtom } from 'modules/yield/state/yieldRawStateAtom'
 
 import { tradeTypeAtom } from './tradeTypeAtom'
 
+import { swapDerivedStateAtom } from '../../swap/state/swapRawStateAtom'
 import { TradeDerivedState, TradeType } from '../types'
 
 export const derivedTradeStateAtom = atom<TradeDerivedState | null>((get) => {
